projects/week-15/requirements.txt
jupyter
pylint
pillow
